Add a README file and some cleanup across.
