+++ README.md
@@ -1,0 +1,34 @@
+# Store Application
+
+Store Application using minio-js library to manage product assets.
+
+- [Prerequisites](#prerequisites)
+- [Installation](#installation)
+- [Running the App](#running-the-app)
+- [Additional links](#additional-links)
+
+## Prerequisites
+
+- Install [nodejs](https://nodejs.org)
+
+## Installation
+
+```bash
+$ git clone https://github.com/minio/minio-js-store-app
+$ cd minio-js-store-app
+$ npm install
+```
+
+## Running the App.
+
+```bash
+$ node minio-store.js
+Express started on http://localhost:3000; press Ctrl-C to terminate.
+```
+
+Now point your browser at [http://localhost:3000](http://localhost:3000).
+
+## Additional links
+- [Minio JavaScript Library for Amazon S3 compatible cloud storage](https://github.com/minio/minio-js)
+- [Minio JavaScript API Reference](https://github.com/minio/minio-js/blob/master/API.md)
+- [More API examples](https://github.com/minio/minio-js#examples)
