+++ package.json
@@ -10,8 +10,7 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.13.4",
-    "express3-handlebars": "^0.5.2",
+    "express-handlebars": "^3.0.0",
     "minio": "^1.0.0"
-   
   }
 }
